`u32` in `legogroth16`

+++ legogroth16/src/data_structures.rs
@@ -66,7 +66,7 @@ pub struct VerifyingKey<E: Pairing> {
     /// The element `eta*gamma^-1 * G` in `E::G1`.
     pub eta_gamma_inv_g1: E::G1Affine,
     /// No of witness to commit
-    pub commit_witness_count: usize,
+    pub commit_witness_count: u32,
 }
 
 /// A verification key in the Groth16 SNARK with CP_link verification parameters
@@ -194,11 +194,11 @@ pub struct LinkPublicGenerators<E: Pairing> {
 }
 
 impl<E: Pairing> VerifyingKey<E> {
-    pub fn num_public_inputs(&self) -> usize {
-        self.gamma_abc_g1.len() - self.commit_witness_count
+    pub fn num_public_inputs(&self) -> u32 {
+        self.gamma_abc_g1.len() as u32 - self.commit_witness_count
     }
 
-    pub fn num_committed_witnesses(&self) -> usize {
+    pub fn num_committed_witnesses(&self) -> u32 {
         self.commit_witness_count
     }
 
@@ -206,8 +206,8 @@ impl<E: Pairing> VerifyingKey<E> {
     pub fn get_commitment_key_for_witnesses(&self) -> Vec<E::G1Affine> {
         let start = self.num_public_inputs();
         let end = start + self.commit_witness_count;
-        let mut key = Vec::with_capacity(self.commit_witness_count + 1);
-        key.extend_from_slice(&self.gamma_abc_g1[start..end]);
+        let mut key = Vec::with_capacity(self.commit_witness_count as usize + 1);
+        key.extend_from_slice(&self.gamma_abc_g1[start as usize..end as usize]);
         key.push(self.eta_gamma_inv_g1);
         key
     }
